add DAO tests for ENTIRE_MUNICIPALITY and Area

+++ server/test_integration/dao.test.ts
@@ -3,6 +3,7 @@ import { db } from "../src/db/dao";
 import { AreaType, KxDocumentType, Scale, Stakeholders } from "../src/models/enum";
 import { KxDocument } from "../src/models/model";
 import { ObjectId } from "mongoose";
+import { KIRUNA_COORDS } from "../src/utils";
 
 const list: ObjectId[] = [];
 const date = new Date();
@@ -112,6 +113,31 @@ describe("Test DAO", () => {
             expect(res2).toBeTruthy();
         }
     });
+
+    test("Test query with whole municipality coordinates", async () => {
+        const res = await db.createKxDocument({
+            title: "test",
+            stakeholders: [],
+            scale_info: Scale.TEXT,
+            scale: 0,
+            issuance_date: date,
+            type: KxDocumentType.INFORMATIVE,
+            connections: 0,
+            language: "Italian",
+            doc_coordinates: {
+                type: AreaType.ENTIRE_MUNICIPALITY,
+            },
+            description: "Test",
+        });
+        if (res && res._id) {
+            expect(res).toHaveProperty("_id");
+            const get = await db.getKxDocumentById(res._id.toString());
+            expect(get?.doc_coordinates).toEqual({type: AreaType.ENTIRE_MUNICIPALITY});
+            const res2 = await db.deleteKxDocument(res._id.toString());
+            expect(res2).toBeTruthy();
+        }
+    });
+
     test("Test query with Point coordinates", async () => {
         const res = await db.createKxDocument({
             title: "test",
@@ -132,6 +158,28 @@ describe("Test DAO", () => {
             expect(res).toHaveProperty("_id");
             const get = await db.getKxDocumentById(res._id.toString());
             expect(get?.doc_coordinates).toEqual({type: AreaType.POINT, coordinates: [20.26, 67.845]});
+            const res2 = await db.deleteKxDocument(res._id.toString());
+            expect(res2).toBeTruthy();
+        }
+    });
+
+    test("Test query with Area (Polygons)", async () => {
+        const res = await db.createKxDocument({
+            title: "test",
+            stakeholders: [],
+            scale_info: Scale.TEXT,
+            scale: 0,
+            issuance_date: date,
+            type: KxDocumentType.INFORMATIVE,
+            connections: 0,
+            language: "Italian",
+            doc_coordinates: { type: AreaType.AREA, coordinates: [[KIRUNA_COORDS, KIRUNA_COORDS.map(c => c + 0.5), KIRUNA_COORDS.map(c => c - 0.5)]] },
+            description: "Test",
+        });
+        if (res && res._id) {
+            expect(res).toHaveProperty("_id");
+            const get = await db.getKxDocumentById(res._id.toString());
+            expect(get?.doc_coordinates).toEqual({ type: AreaType.AREA, coordinates: [[KIRUNA_COORDS, KIRUNA_COORDS.map(c => c + 0.5), KIRUNA_COORDS.map(c => c - 0.5)]] });
             const res2 = await db.deleteKxDocument(res._id.toString());
             expect(res2).toBeTruthy();
         }
